mqttmanager/requirements.txt
flask
pymongo
pandas
wtforms
jsonpickle
flask-login
flask-wtf
flask-cors
pyopenssl
paho-mqtt